Everythings work

+++ app.py
@@ -5,14 +5,19 @@ import bcrypt
 import os
 from werkzeug.utils import secure_filename
 from uuid import uuid4
+from datetime import datetime
+import string, secrets
 
 UPLOAD_FOLDER = 'static/uploads'
 ALLOWED_EXTENSIONS = {'png', 'jpg', 'jpeg', 'gif'}
 
 app = Flask(__name__)
-app.config.from_pyfile('config.py')
 app.secret_key = token_hex(16)
 app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
+app.config['MYSQL_HOST']='localhost'
+app.config['MYSQL_USER']='root'
+app.config['MYSQL_PASSWORD']=''
+app.config['MYSQL_DB']='db_voting'
 
 def allowed_file(filename):
     return '.' in filename and filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS
@@ -20,10 +25,7 @@ def allowed_file(filename):
 
 mysql = MySQL(app)
 
-def login_condition():
-    if 'id_admin' not in session:
-        return redirect(url_for('login'))
-    return None
+
 
 @app.route('/')
 def index():
@@ -55,11 +57,10 @@ def login():
     
     return render_template('login.html')
 
-
 @app.route('/dashboard')
-def dashboard():
-    # if 'id_admin' not in session:
-    #     return redirect(url_for('login'))
+def dashboard(pesan=None):
+    if 'id_admin' not in session:
+        return redirect(url_for('login'))
     cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
     cursor.execute('SELECT * FROM pemilihan')
     pemilihan = cursor.fetchall()
@@ -67,10 +68,9 @@ def dashboard():
     return render_template('pemilihan/index.html', data=pemilihan)
 
 @app.route('/tambah_pemilihan', methods=['GET','POST'])
-
 def tambah_pemilihan():
-    # if 'id_admin' not in session:
-    #     return redirect(url_for('login'))
+    if 'id_admin' not in session:
+        return redirect(url_for('login'))
     
     if request.method == 'POST':
         nama_pemilihan = request.form['nama_pemilihan']
@@ -90,8 +90,8 @@ def tambah_pemilihan():
 
 @app.route('/edit_pemilihan/<int:id>', methods=['GET','POST'])
 def edit_daftar_pemilihan(id):
-    # if 'id_admin' not in session:
-    #     return redirect(url_for('login'))
+    if 'id_admin' not in session:
+        return redirect(url_for('login'))
     
     cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
     cursor.execute('SELECT * FROM pemilihan WHERE id_pemilihan = %s', (id,))
@@ -116,20 +116,57 @@ def edit_daftar_pemilihan(id):
 @app.route('/verify', methods=['GET','POST'])
 def verify():
     if request.method == 'POST':
-        code = request.form['code']
+        code = request.form['kode_verifikasi']
         
-        cursor = mysql.connection.cursor()
-        cursor.execute('SELECT * FROM verification_codes WHERE code = %s', (code,))
+        cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
+        cursor.execute('SELECT * FROM voting WHERE verification_code = %s', (code,))
         verification = cursor.fetchone()
         cursor.close()
-        
+
+        if verification['id_candidate'] is not None:
+            return render_template('verify.html', error='This verification code has already been used to vote.')        
+
+        cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
+        cursor.execute('SELECT pemilihan.nama_pemilihan FROM pemilihan WHERE pemilihan.id_pemilihan =%s', (verification['id_pemilihan'],))
+        nama_pemilihan = cursor.fetchone()
+        cursor.close()
+
+        cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
+        cursor.execute('SELECT * FROM candidates WHERE candidates.id_pemilihan =%s', (verification['id_pemilihan'],))
+        kandidat = cursor.fetchall()
+        cursor.close()
+
+        cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
+        cursor.execute('SELECT * FROM voters WHERE voters.id_voter =%s', (verification['id_voter'],))
+        voter = cursor.fetchone()
+        cursor.close()
+
         if verification:
             session['verified'] = True
-            return redirect(url_for('dashboard'))
+            return render_template('voting.html',data=kandidat, pemilihan=nama_pemilihan, id_voting=verification['id_voting'],nama_voter=voter['nama'])
         else:
             return render_template('verify.html', error='Invalid verification code')
     
     return render_template('verify.html')
+
+@app.route('/voted', methods=['POST'])
+def voted():
+    if 'verified' not in session:
+        return redirect(url_for('verify'))
+    
+    id_voting=request.form['id_voting']
+    id_candidate=request.form['id_candidate']
+
+    cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
+
+    
+    cursor.execute('UPDATE voting SET id_candidate = %s, updated_at=%s WHERE id_voting = %s', (id_candidate, datetime.now(), id_voting))
+    mysql.connection.commit()
+    cursor.close()
+    
+    session.pop('verified', None)
+    
+    return render_template('voted.html')
 
 def cek_koneksi():
     try:
@@ -163,7 +200,6 @@ def logout():
     session.clear()
     return redirect(url_for('index'))
 
-
 @app.route('/kelas')
 def kelas():
     cursor= mysql.connection.cursor(MySQLdb.cursors.DictCursor)
@@ -171,7 +207,6 @@ def kelas():
     kelas= cursor.fetchall()
     cursor.close()
     return render_template('kelas/index.html', data=kelas)
-
 
 @app.route('/tambah_kelas',methods=['GET','POST'])
 def tambah_kelas():
@@ -255,19 +290,20 @@ def hapus_voter(id):
     cursor.close()
     return redirect(url_for('voters'))
 
-
 @app.route('/kandidat')
 def kandidat(): 
+    if 'id_admin' not in session:
+        return redirect(url_for('login'))
     cursor= mysql.connection.cursor(MySQLdb.cursors.DictCursor)
     cursor.execute('SELECT * FROM candidates JOIN pemilihan ON candidates.id_pemilihan = pemilihan.id_pemilihan WHERE pemilihan.status="T" ')
     kandidat= cursor.fetchall()
     cursor.close()
     return render_template('kandidat/index.html', data=kandidat)
 
-
-
 @app.route('/tambah_kandidat',methods=['GET','POST'])
 def tambah_kandidat():
+    if 'id_admin' not in session:
+        return redirect(url_for('login'))
     cursor= mysql.connection.cursor(MySQLdb.cursors.DictCursor)
     cursor.execute('SELECT * FROM pemilihan WHERE status="T" ')
     pemilihan= cursor.fetchall()
@@ -296,6 +332,8 @@ def tambah_kandidat():
 
 @app.route('/delete_kandidat/<int:id>',methods=['POST'])
 def hapus_kandidat(id):
+    if 'id_admin' not in session:
+        return redirect(url_for('login'))
     if request.method=='POST':
         cursor=mysql.connection.cursor(MySQLdb.cursors.DictCursor)
         cursor.execute('DELETE FROM candidates WHERE id_candidate=%s',[id])
@@ -305,6 +343,8 @@ def hapus_kandidat(id):
     
 @app.route('/edit_kandidat/<int:id>',methods=['GET','POST'])
 def edit_kandidat(id):
+    if 'id_admin' not in session:
+        return redirect(url_for('login'))
     cursor= mysql.connection.cursor(MySQLdb.cursors.DictCursor)
     cursor.execute('SELECT * FROM pemilihan WHERE status="T" ')
     pemilihan= cursor.fetchall()
@@ -333,5 +373,93 @@ def edit_kandidat(id):
     
     return render_template('kandidat/edit.html',data=kandidat, pemilihan=pemilihan)
 
+def generate_verification_code(length=6):
+    characters = string.ascii_uppercase + string.digits  # A-Z dan 0-9
+    code = ''
+    for i in range(length):
+        random_char = secrets.choice(characters)
+        code += random_char
+    return code
+
+@app.route('/generate_validation_code/<int:id>')
+def generate_validation_code(id):
+
+
+    if 'id_admin' not in session:
+        return redirect(url_for('login'))
+
+    cursor=mysql.connection.cursor(MySQLdb.cursors.DictCursor)
+    cursor.execute("SELECT * FROM voters")
+    voters=cursor.fetchall()
+    if len(voters) == 0:
+        return "No voters found. Please add voters before generating a verification code."
+    cursor.close()
+
+    for voter in voters:
+        id_voter=voter['id_voter']
+        id_pemilihan=id
+        created_at = datetime.now()
+        verification_code=generate_verification_code(5)
+        # cek apakah verification code sudah ada di database
+        cursor=mysql.connection.cursor(MySQLdb.cursors.DictCursor)  
+        cursor.execute('SELECT * FROM voting WHERE verification_code=%s', [verification_code])
+        existing_code=cursor.fetchone()
+        cursor.close()
+        if existing_code is not None:
+            # jika sudah ada, generate ulang
+            verification_code=generate_verification_code(5)
+        
+         # simpan ke database
+         #cek apakah user sudah memiliki kode verifikasi untuk pemilihan ini
+        cursor=mysql.connection.cursor(MySQLdb.cursors.DictCursor)
+        cursor.execute('SELECT * FROM voting WHERE id_voter=%s AND id_pemilihan=%s', (id_voter, id_pemilihan))
+        existing_voting=cursor.fetchone()
+        cursor.close()
+        if existing_voting is not None:
+            # jika sudah ada, lewati
+            continue
+        
+         #simpan kode verifikasi baru
+        cursor = mysql.connection.cursor()
+        cursor.execute('INSERT INTO voting (verification_code,id_voter, id_pemilihan, created_at,updated_at) VALUES (%s, %s, %s, %s,%s)', (verification_code, id_voter, id_pemilihan, created_at,created_at))
+        mysql.connection.commit()
+        cursor.close()
+    
+    return redirect(url_for('dashboard', pesan="Verification codes generated successfully."))
+    
+
+@app.route('/cetak_validation_code/<int:id>')
+def cetak_validation_code(id):
+    if 'id_admin' not in session:
+        return redirect(url_for('login'))
+    cursor=mysql.connection.cursor(MySQLdb.cursors.DictCursor)
+    cursor.execute('SELECT voting.verification_code, voters.nama AS nama_voter,pemilihan.nama_pemilihan, kelas.kode_kelas ' \
+    'FROM voting ' \
+    'JOIN voters ON voting.id_voter = voters.id_voter ' \
+    'JOIN pemilihan ON pemilihan.id_pemilihan = voting.id_pemilihan ' \
+    'JOIN kelas ON voters.id_kelas = kelas.id_kelas WHERE voting.id_pemilihan=%s ', [id] )
+    codes=cursor.fetchall()
+    cursor.close()
+    return render_template('/pemilihan/voting_codes.html', data=codes)
+
+
+@app.route('/hasil_pemilihan/<int:id>')
+def hasil_pemilihan(id):
+    if 'id_admin' not in session:
+        return redirect(url_for('login'))
+    total=0
+    cursor=mysql.connection.cursor(MySQLdb.cursors.DictCursor)
+    cursor.execute('SELECT candidates.nama,(SELECT COUNT(*) FROM voting WHERE voting.id_candidate=candidates.id_candidate) as jumlah_voting FROM pemilihan JOIN candidates ON candidates.id_pemilihan=pemilihan.id_pemilihan WHERE pemilihan.id_pemilihan=%s ', [id] )
+
+    hasil=cursor.fetchall()
+ 
+    cursor.close()
+    for item in hasil:
+        total=sum(row['jumlah_voting'] for row in hasil)
+
+
+    print (hasil)
+
+    return render_template('hasil.html', data=hasil, total=total)
 if __name__ == '__main__':
     app.run(debug=True)
